Add more validation to service. Move service validation functions to service package.
