+++ protobuf/types/service.proto
@@ -15,7 +15,7 @@ message Service {
   message Event {
     // Event's key.
     string key = 4 [
-      (gogoproto.moretags) = 'hash:"name:4" validate:"printascii"'
+      (gogoproto.moretags) = 'hash:"name:4" validate:"required,printascii"'
     ];
 
     // Event's name.
@@ -38,7 +38,7 @@ message Service {
   message Task {
     // Task's key.
     string key = 8 [
-      (gogoproto.moretags) = 'hash:"name:8" validate:"printascii"'
+      (gogoproto.moretags) = 'hash:"name:8" validate:"required,printascii"'
     ];
 
     // Task's name.
@@ -64,10 +64,9 @@ message Service {
 
   // Parameter describes the task's inputs, the task's outputs, and the event's data.
   message Parameter {
-
     // Parameter's key.
     string key = 8 [
-      (gogoproto.moretags) = 'hash:"name:8" validate:"printascii"'
+      (gogoproto.moretags) = 'hash:"name:8" validate:"required,printascii"'
     ];
 
     // Parameter's name.
@@ -138,12 +137,12 @@ message Service {
   message Dependency {
     // Dependency's key.
     string key = 8 [
-      (gogoproto.moretags) = 'hash:"name:8" validate:"printascii"'
+      (gogoproto.moretags) = 'hash:"name:8" validate:"required,printascii"'
     ];
 
     // Image's name of the container.
     string image = 1 [
-      (gogoproto.moretags) = 'hash:"name:1" validate:"printascii"'
+      (gogoproto.moretags) = 'hash:"name:1" validate:"required,printascii"'
     ];
 
     // List of volumes.
@@ -179,7 +178,7 @@ message Service {
 
   // Service's hash.
   bytes hash = 10 [
-    (gogoproto.moretags) = 'hash:"-" validate:"required"',
+    (gogoproto.moretags) = 'hash:"-" validate:"required,hash"',
     (gogoproto.customtype) = "github.com/mesg-foundation/engine/hash.Hash",
     (gogoproto.nullable) = false
   ];
@@ -232,7 +231,7 @@ message Service {
 
   // The address of the service.
   bytes address = 14 [
-    (gogoproto.moretags) = 'hash:"-" validate:"accaddress"',
+    (gogoproto.moretags) = 'hash:"-" validate:"required,accaddress"',
     (gogoproto.casttype) = "github.com/cosmos/cosmos-sdk/types.AccAddress"
   ];
 }
